+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage/Controllers/InstancesController.cs
@@ -895,55 +895,6 @@ namespace Altinn.Platform.Storage.Controllers
         }
 
         /// <summary>
-        /// Restores a soft deleted instance
-        /// </summary>
-        /// <param name="instanceOwnerId">instance owner</param>
-        /// <param name="instanceGuid">instance id</param>
-        /// <returns>True if the instance was restored.</returns>
-        [HttpPut("{instanceOwnerId:int}/{instanceGuid:guid}/restore")]
-        public async Task<ActionResult> Restore(int instanceOwnerId, Guid instanceGuid)
-        {
-            string instanceId = $"{instanceOwnerId}/{instanceGuid}";
-
-            Instance instance;
-
-            try
-            {
-                instance = await _instanceRepository.GetOne(instanceId, instanceOwnerId);
-            }
-            catch (DocumentClientException dce)
-            {
-                if (dce.Error.Code.Equals("NotFound"))
-                {
-                    return NotFound($"Didn't find the object that should be restored with instanceId={instanceId}");
-                }
-
-                return StatusCode(500, $"Unknown database exception in restore: {dce}");
-            }
-
-            if (instance.InstanceState.IsMarkedForHardDelete)
-            {
-                return BadRequest("Instance was permanently deleted and cannot be restored.");
-            }
-            else if (instance.InstanceState.IsDeleted)
-            {
-                instance.InstanceState.IsDeleted = false;
-
-                try
-                {
-                    await _instanceRepository.Update(instance);
-                    return Ok(true);
-                }
-                catch (Exception e)
-                {
-                    return StatusCode(500, $"Unknown exception in restore: {e}");
-                }
-            }
-
-            return Ok(true);
-        }
-
-        /// <summary>
         /// Delete an instance
         /// </summary>
         /// <param name="instanceGuid">instance id</param>
